requirements.txt
accelerate==1.10.1
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
apsw==3.51.0.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorlog==6.10.1
dill==0.4.0
distro==1.9.0
docling==2.63.0
docling-core==2.52.0
docling-ibm-models==3.10.2
docling-parse==4.7.1
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Faker==37.12.0
fastapi==0.122.0
filelock==3.19.1
filetype==1.2.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonlines==4.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
latex2mathml==3.78.1
lxml==6.0.2
markdown-it-py==3.0.0
marko==2.2.1
MarkupSafe==3.0.3
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.18
networkx==3.2.1
numpy==2.0.2
ocrmac==1.0.0
ollama==0.6.1
omegaconf==2.3.0
openai==2.8.1
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.1.0
psutil==7.1.3
pyclipper==1.3.0.post6
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.19.2
pylatexenc==2.10
pyobjc-core==12.0
pyobjc-framework-Cocoa==12.0
pyobjc-framework-CoreML==12.0
pyobjc-framework-Quartz==12.0
pyobjc-framework-Vision==12.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.4.2
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
rtree==1.4.1
safetensors==0.7.0
scipy==1.13.1
semchunk==2.2.2
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlite-vec==0.1.6
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.3
tree-sitter==0.23.2
tree-sitter-c==0.23.6
tree-sitter-java==0.23.5
tree-sitter-javascript==0.23.1
tree-sitter-python==0.23.6
tree-sitter-typescript==0.23.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uv==0.9.13
uvicorn==0.38.0
xlsxwriter==3.2.9
